requirements.txt
backoff==2.2.1
djangorestframework==3.15.2
functions_framework==3.8.1

google-cloud-tasks==2.16.5
google-cloud-secret-manager==2.20.2
google-crc32c==1.5.0

requests==2.32.3
ratelimit==2.2.1

growattServer==1.7.1

pandas
openpyxl